src/openwebui/requirements.txt
fastapi~=0.127.0
uvicorn[standard]~=0.34.0
pydantic~=2.12.0
python-dotenv~=1.0.1
openai~=1.16.2
qdrant-client~=1.16.0
langfuse~=2.52.0
azure-identity~=1.17.1
azure-keyvault-secrets~=4.8.0

# LlamaIndex - install core and let dependencies resolve
llama-index-core~=0.14.10
llama-index-embeddings-azure-openai
llama-index-llms-azure-openai
llama-index-vector-stores-qdrant
